Listening to Sonar cloud

+++ migrations-data/updateORC3883.js
@@ -33,17 +33,13 @@ async function deleteOldPark(oldORC){
   const parkData = await getPark(oldORC);
   if (parkData.length > 0) {
     const subAreaList = await getSubAreas(oldORC);
-    for (area of subAreaList){
+    for (const area of subAreaList){
       const activityList = area.activities.values;
       for (const activity of activityList){
         const recordsList = await getActivityRecords(area, activity);
         for (const record of recordsList) {
           const varianceList = await getVariances(record, oldORC);
-          if(varianceList.length > 0){
-            for (const variance of varianceList){
-              await deleteVariance(variance, record.date, record.pk);
-            }
-          }
+          await deleteVariance(varianceList, record.pk);
         }
       }
       await deleteSubArea(area.pk, area.sk);
@@ -66,7 +62,7 @@ async function deletePark(oldORC){
       },
       ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)',
     };
-    const res = await dynamodb.deleteItem(deletePark).promise();
+    await dynamodb.deleteItem(deletePark).promise();
   }catch (error){
     console.log("Failed to delete park ", oldORC);
     console.log(error);
@@ -96,27 +92,33 @@ async function deleteSubArea(pk, sk){
       },
       ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)',
     };
-    const res = await dynamodb.deleteItem(deleteSubArea).promise();
+    await dynamodb.deleteItem(deleteSubArea).promise();
   }catch (error){
     console.log("Failed to delete sub area: ", pk, " ", sk);
     console.log(error);
   }
 }
 
-async function deleteVariance(variance, date, sk){
+async function deleteVariance(varianceList, recordPK){
   try{
-    console.log("Deleting Variance: ", variance.pk);
-    const deleteVariance = {
-      TableName: TABLE_NAME,
-      Key: {
-        pk: {S: `variance::${variance.orcs}::${date}`},
-        sk: {S: `${sk}`}
-      },
-      ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)',
-    };
-    const res = await dynamodb.deleteItem(deleteVariance).promise();
+    if(varianceList.length > 0){
+      for(const variance of varianceList){
+        console.log("Deleting Variance: ", variance.pk);
+        const deleteVariance = {
+          TableName: TABLE_NAME,
+          Key: {
+            pk: {S: variance.pk},
+            sk: {S: variance.sk}
+          },
+          ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)',
+        };
+        await dynamodb.deleteItem(deleteVariance).promise();
+      }
+    } else {
+      console.log("There are no variances to delete for record: ", recordPK)
+    }
   }catch(error){
-    console.log("Failed to delete variance for: ", variance.pk);
+    console.log("Failed to delete variance for: ", recordPK);
     console.log(error);
   }
 }
@@ -133,29 +135,23 @@ async function convertPark(oldORC, newORC, newParkName){
     await createNewPark(newORC, newParkName);
     const subAreaList = await getSubAreas(oldORC);
     console.log("Creating the new subareas for: ", newORC);
-    for (const area of subAreaList){
-     await addSubArea(newORC, area, newParkName);
-    }
+    await addSubAreas(subAreaList, newORC, newParkName);
+    
     for(const area of subAreaList){ 
       console.log("Checking For Configs: ", area.pk, ":", area.sk); 
       const activityList = area.activities.values
       for(const activity of activityList){
-        const config = await getConfigs(area);
-        //Check length to work with legacy Records   
-        if(config.length > 0){
-          if(config[0].parkName != newParkName){
-            console.log("Updating config: ", activity, " for ", area.subAreaName);
-            await updateConfig(config, newORC, newParkName);
-          }
+        const config = await getConfigs(area); 
+        if(config.parkName && config.parkName != newParkName){
+          console.log("Updating config: ", activity, " for ", area.subAreaName);
+          await updateConfig(config, newORC, newParkName);
         }
       }
       for (const activity of activityList){
         const recordsList = await getActivityRecords(area, activity);
         for (const record of recordsList) {
-          const varianceList = await getVariances(record, oldORC); // Access each object using recordsArray[record]
-          for (const variance of varianceList){
-            await createVariance(variance, record.date, newORC, newParkName);
-          }
+          const varianceList = await getVariances(record, oldORC);
+          await createVariance(varianceList, record.date, newORC, newParkName);
           await updateRecord(record, newORC);
         }
       }
@@ -166,29 +162,31 @@ async function convertPark(oldORC, newORC, newParkName){
   }
 }
 
-async function createVariance(variance, date, newORC, newParkName){
+async function createVariance(varianceList, date, newORC, newParkName){
   try{
-    console.log("Re-creating variance: ", variance.pk);
-    const newVariance = {
-      TableName: TABLE_NAME,
-      ConditionExpression: "attribute_not_exists(pk)",
-      Item: {
-          pk: { S: `variance::${newORC}::${date}` },
-          sk: { S: `${variance.sk}` },
-          orcs: { S: `${newORC}` },
-          parkName: { S: `${newParkName}` },
-          roles: { L: [{S: "sysadmin"}, {S: `${newORC}:${variance.subAreaId}`}] },
-          bundle: { S: `${variance.bundle}`},
-          subAreaName:{ S: variance.subAreaName },
-          subAreaId: { S: variance.subAreaId },
-          notes: { S: variance.notes },
-          resolved: {BOOL: variance.resolved },
-      },
+    for(const variance of varianceList){
+      console.log("Re-creating variance: ", variance.pk);
+      const newVariance = {
+        TableName: TABLE_NAME,
+        ConditionExpression: "attribute_not_exists(pk)",
+        Item: {
+            pk: { S: `variance::${newORC}::${date}` },
+            sk: { S: `${variance.sk}` },
+            orcs: { S: `${newORC}` },
+            parkName: { S: `${newParkName}` },
+            roles: { L: [{S: "sysadmin"}, {S: `${newORC}:${variance.subAreaId}`}] },
+            bundle: { S: `${variance.bundle}`},
+            subAreaName:{ S: variance.subAreaName },
+            subAreaId: { S: variance.subAreaId },
+            notes: { S: variance.notes },
+            resolved: {BOOL: variance.resolved },
+        },
+      }
+      if (variance.fields && variance.fields.length != 0) {
+        newVariance.Item.fields = { SS: variance.fields };
+      }
+      await dynamodb.putItem(newVariance).promise();
     }
-    if (variance.fields && variance.fields.length != 0) {
-      newVariance.Item.fields = { SS: variance.fields };
-    }
-    const res = await dynamodb.putItem(newVariance).promise();
   } catch (error){
     if (error.code === "ConditionalCheckFailedException") {
       console.log("Variance already exists: ", variance.pk);
@@ -219,7 +217,7 @@ async function updateRecord(record, newORC){
       ReturnValues: 'ALL_NEW', 
       ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)' 
     };    
-    const { Attributes } = await dynamodb.updateItem(updateRecord).promise();
+    await dynamodb.updateItem(updateRecord).promise();
   } catch (error) {
     console.log('Error updating item:', error);
   }
@@ -241,7 +239,6 @@ async function updateConfig(config, newORC, newParkName){
         
         UpdateExpression: 'set #orcs = :newOrcs, #parkName =:newParkName',
         ReturnValues: 'ALL_NEW',
-        TableName: TABLE_NAME,
         ConditionExpression: 'attribute_exists(pk) AND attribute_exists(sk)',
       };
       const { Attributes } = await dynamodb.updateItem(updateConfig).promise();
@@ -268,7 +265,7 @@ async function createNewPark(newORC, newParkName){
             subAreas: { L: [] },
         },
     };
-    const res = await dynamodb.putItem(newPark).promise();
+    await dynamodb.putItem(newPark).promise();
   } catch (error) {
       if (error.code === "ConditionalCheckFailedException") {
           console.log("Park already exists");
@@ -279,45 +276,47 @@ async function createNewPark(newORC, newParkName){
   }
 }
 
-async function addSubArea(newORC, area, newParkName){
+async function addSubAreas(oldSubAreaList, newORC, newParkName){
   try {
-      // This needs to makesure the ORC doesnt already contain the subarea.
+      // This needs to makesure the newORC doesnt already have the subareas.
       // Once that is confirmed create the object
       // Add info to the park
       // Create the actual subarea
       const newSubAreaList = await getSubAreas(newORC);
-      if (newSubAreaList.some(obj => obj.sk === area.sk)) {
-        console.log('The subarea ', area.sk, " already exists for ", newORC);
-      } else { 
-          console.log("Creating subAreaName: ", area.subAreaName, " for ", newORC);
-          const newSubAreaOBJ = {
-              activities: area.activities.values,
-              orcs: newORC,
-              managementArea: area.managementArea || " ",
-              section: area.section,
-              region: area.region,
-              bundle: area.bundle,
-              subAreaName: area.subAreaName
-          };
-          let subAreaObj = getValidSubareaObj(newSubAreaOBJ, newParkName);
-          let transactionObj = { TransactItems: [] };
-          transactionObj.TransactItems.push({
-              Update: createUpdateParkWithNewSubAreaObj(
-                  subAreaObj.subAreaName,
-                  area.sk,
-                  subAreaObj.isLegacy,
-                  subAreaObj.orcs
-              ),
-          });
-          transactionObj.TransactItems.push({
-              Put: createPutSubAreaObj(subAreaObj, area.sk, newParkName),
-          });
-          const res = await dynamodb.transactWriteItems(transactionObj).promise();
+      for (const area of oldSubAreaList){
+        if (newSubAreaList.some(obj => obj.sk === area.sk)) {
+          console.log('The subarea ', area.sk, " already exists for ", newORC);
+          continue;
+        }
+        console.log("Creating subAreaName: ", area.subAreaName, " for ", newORC);
+        const newSubAreaOBJ = {
+            activities: area.activities.values,
+            orcs: newORC,
+            managementArea: area.managementArea || " ",
+            section: area.section,
+            region: area.region,
+            bundle: area.bundle,
+            subAreaName: area.subAreaName
+        };
+        let subAreaObj = getValidSubareaObj(newSubAreaOBJ, newParkName);
+        let transactionObj = { TransactItems: [] };
+        transactionObj.TransactItems.push({
+            Update: createUpdateParkWithNewSubAreaObj(
+                subAreaObj.subAreaName,
+                area.sk,
+                subAreaObj.isLegacy,
+                subAreaObj.orcs
+            ),
+        });
+        transactionObj.TransactItems.push({
+            Put: createPutSubAreaObj(subAreaObj, area.sk, newParkName),
+        });
+        await dynamodb.transactWriteItems(transactionObj).promise();
       }
-  } catch (error) {
+    } catch (error) {
       console.error("Error adding subarea:", error);
+    }
   }
-}
     
 async function getConfigs(subArea) {
   const subAreaID = subArea.sk;
